// src/core/enums/material-uses.enum.ts
export enum MaterialUsesEnum {
  ADD_LIFE = 'Restaura vida',
  SELL = 'Se puede vender',
  IMPROVE_ARMOR = 'Mejora algunas armaduras',
  ELIXIR = 'Ingrediente de elixir',
  DEFENCE_COOKING = 'Aumenta la defensa cuando se cocina',
  ATACK_COOKING = 'Aumenta el poder de ataque cuando se cocina',
  HEARTS_COOKING = 'Aumenta el máximo de corazones cuando se cocina',
  ANCESTRAL_WEAPONS = 'Elaboración de armas ancestrales',
  SPECIFIC_RECIPES = 'Cocina general y recetas específicas',
  HEAT_RESISTANCE_COOKING = 'Resistencia al calor cuando se cocina',
  RESISTANCE_COOKING = 'Resistencia cuando se cocina',
  CHAMPION_WEAPONS = 'Elaboración de armas de campeón',
  RESISTANCE_HORSE = 'Resistencia para el caballo',
  RESISTANCE_COLD = 'Resistencia al frio cuando se cocina'
}